Int generator + fps counter

+++ csharp/generators/DataGenerators.cs
@@ -2,9 +2,9 @@ namespace DataGeneration
 {
     public class IntGenerator: AbstractDataGenerator<int>
     {
-        public IntGenerator(int hz): base(hz) {}
+        public IntGenerator(double hz): base(hz) {}
 
-        public IntGenerator(int hz, int seed): base(hz, seed) {}
+        public IntGenerator(double hz, int seed): base(hz, seed) {}
         
         public override int GenerateNewData()
         {
